# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "My Api",
        "description": "Lumen Swagger Api",
        "license": {
            "name": "my api"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api/v1",
            "description": "Api Documentations my api"
        }
    ],
    "paths": {
...
        "/resourceType": {
            "get": {
                "tags": [
                    "ResourceType"
                ],
                "summary": "List resource type",
                "description": "Returns data of resource types",
                "operationId": "listResourceType",
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "authorization": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "ResourceType"
                ],
                "summary": "Store resource type",
                "operationId": "storeResourceType",
                "requestBody": {
                    "description": "Input data format",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "description"
                                ],
                                "properties": {
                                    "name": {
                                        "description": "Nombre de el tipo de recurso",
                                        "type": "string"
                                    },
                                    "description": {
                                        "description": "Descripcion del tipo de recurso",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "authorization": []
                    }
                ]
            }
        },
        "/resourceType/{id}": {
            "get": {
                "tags": [
                    "ResourceType"
                ],
                "summary": "Show resource type",
                "description": "Returns data of resource type",
                "operationId": "showResourceType",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "resourceType id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "authorization": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "ResourceType"
                ],
                "summary": "Update resource type",
                "operationId": "updateResourceType",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Resource type id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Input data format",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "description"
                                ],
                                "properties": {
                                    "name": {
                                        "description": "Nombre de el tipo de recurso",
                                        "type": "string"
                                    },
                                    "description": {
                                        "description": "Descripcion del tipo de recurso",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "authorization": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "ResourceType"
                ],
                "summary": "Delete existing resource type",
                "description": "Deletes a record and returns no content",
                "operationId": "deleteResourceType",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "resource type id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Resource Type Not Found"
                    }
                },
                "security": [
                    {
                        "authorization": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "authorization": {
                "type": "apiKey",
                "name": "authorization",
                "in": "header"
            }
        }
    }
}